+++ src/config/args/tests/options/check_etag.rs
@@ -149,7 +149,7 @@ mod tests {
     }
 
     #[test]
-    fn with_custom_value_with_conflict_express_onezone() {
+    fn with_custom_value_with_conflict_source_express_onezone() {
         init_dummy_tracing_subscriber();
 
         let args = vec![
@@ -161,6 +161,26 @@ mod tests {
             "--check-etag",
             "s3://bucket-base-name--usw2-az1--x-s3/source_key",
             "s3://target-bucket/target_key",
+        ];
+
+        let config = build_config_from_args(args);
+
+        assert!(config.is_err());
+    }
+
+    #[test]
+    fn with_custom_value_with_conflict_target_express_onezone() {
+        init_dummy_tracing_subscriber();
+
+        let args = vec![
+            "s3sync",
+            "--source-profile",
+            "source_profile",
+            "--target-profile",
+            "target_profile",
+            "--check-etag",
+            "s3://source-bucket/source_key",
+            "s3://bucket-base-name--usw2-az1--x-s3/target_key",
         ];
 
         let config = build_config_from_args(args);