+++ src/config/args/tests/options/http_proxy.rs
@@ -71,6 +71,27 @@ mod tests {
         assert!(config.is_err());
     }
 
+    #[test]
+    fn with_custom_value_http_auth_error() {
+        init_dummy_tracing_subscriber();
+
+        let args = vec![
+            "s3sync",
+            "--source-profile",
+            "source_profile",
+            "--target-profile",
+            "target_profile",
+            "--http-proxy",
+            "http://user:password@localhost:8080",
+            "s3://source-bucket/source_key",
+            "s3://target-bucket/target_key",
+        ];
+
+        let config = build_config_from_args(args);
+
+        assert!(config.is_err());
+    }
+
     fn init_dummy_tracing_subscriber() {
         let _ = tracing_subscriber::fmt()
             .with_env_filter("dummy=trace")