+++ src/config/args/tests/options/https_proxy.rs
@@ -51,6 +51,34 @@ mod tests {
     }
 
     #[test]
+    fn with_custom_value_auth() {
+        init_dummy_tracing_subscriber();
+
+        let args = vec![
+            "s3sync",
+            "--source-profile",
+            "source_profile",
+            "--target-profile",
+            "target_profile",
+            "--https-proxy",
+            "http://user:password@localhost:8080",
+            "s3://source-bucket/source_key",
+            "s3://target-bucket/target_key",
+        ];
+
+        let config = build_config_from_args(args).unwrap();
+
+        assert_eq!(
+            config.source_client_config.unwrap().https_proxy.unwrap(),
+            "http://user:password@localhost:8080".to_string()
+        );
+        assert_eq!(
+            config.target_client_config.unwrap().https_proxy.unwrap(),
+            "http://user:password@localhost:8080".to_string()
+        );
+    }
+
+    #[test]
     fn with_custom_value_error() {
         init_dummy_tracing_subscriber();