+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2024-06-02
+
+### Added
+
+- Added HTTPS proxy authentication support.
+- Added `--check-etag` and `--check-etag` & `--auto-chunksize` option.
+
+### Changed
+
+- aws-sdk-s3 = "1.32.0"
+
 ## [1.3.0] - 2024-04-28
 
 ### Added